% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature.R
\name{get_signature}
\alias{get_signature}
\title{get_signature}
\usage{
get_signature(rawCounts, groupA, groupB, remove = FALSE)
}
\arguments{
\item{rawCounts}{The RNA-Seq raw counts expression matrix of related dataset. Each row represents a gene and each
column represents a sample.}

\item{groupA}{Column indexes of the first group samples. The contrast of two groups in DESeq2 is groupA vs groupB.}

\item{groupB}{Column indexes of the second group samples.}

\item{remove}{Logical variable that indicate whether the genes with missing adjusted p-values in DESeq2 result should be removed.}
}
\description{
Compute the gene signature using the current tools in DESeq2 package.
}
