% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_Finding.R, R/Reg_Finding_original.R
\name{Onetailed}
\alias{Onetailed}
\title{Onetailed: Absolute gene permuting GSEA code.}
\usage{
Onetailed(regulon, tvalue, min, max, nPerm)

Onetailed(regulon, tvalue, min, max, nPerm)
}
\description{
Onetailed: Absolute gene permuting GSEA code.

Onetailed: Absolute gene permuting GSEA code.
}
