% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_Finding.R
\name{Reg_Finding}
\alias{Reg_Finding}
\title{Reg_Finding}
\usage{
Reg_Finding(signature, regulon, minsize = 10, maxsize = 1000,
  Permutation = 1000, cutoff = 0.01, sort = TRUE,
  outfile = "Outputs.txt")
}
\arguments{
\item{signature}{Gene expression profile that each gene has a corresponding score. Any computational method is applicable for obtaining this vector.}

\item{regulon}{Regulon object contains the information of transcription factor and target genes, which can be
generated from networks reverse engineered with the ARACNe algorithm.}

\item{minsize}{The minimum size (number of target genes) of output regulon. Only regulons with size larger than
\code{minsize} are considered. The default value is 10.}

\item{maxsize}{The maximum size (number of target genes) of output regulon. Only regulons with size less than
\code{maxnsize} are considered. The default value is 1000.}

\item{Permutation}{The times of gene-permutation in GSEA algorithm. The default value is 1000.}

\item{cutoff}{The cutoff of FDR for selecting the significant TFs. The default value is 0.01.}

\item{sort}{Locaical variable indicate whether the final output result should be sorted in FDR ascending order.}

\item{outfile}{The filename of output result. The file is saved at the current working directory. If \code{outfile == NULL}, algorithm would not
save the final result automatically.}
}
\value{
This function will return a \code{data.frame} with the following components (columns):
  \item{TF_Name}{The gene symbol of the output transcription factor.}
  \item{Positive}{The number of positive target genes of current TF.}
  \item{Negative}{The number of negative target genes of current TF.}
  \item{Target}{The total number of target genes of current TF.}
  \item{Reg_Score}{The regulon score computed by Reg_Finding. See details for more description.}
  \item{FDR}{The FDR value of absolute-based GSEA.}
  \item{Direction}{The indicator to show whether the significant TF is \code{ACTIVATED} or \code{REPRESSED}.}
}
\description{
Reg_Finding is a novel proposed RNA-Seq tool, which can identify the key regulons that are most responsible for the difference in
two group samples. Reg_Finding is especially designed for RNA-Seq data with small sample replicates.
}
\details{
The core idea of identifying the key regulon
is using the enrichment analysis on the (positive/negative) target genes of a given TF. After obtains the gene signature, two types of GSEA
analyses are executed. The first kind of GSEA is reverse positive/negative target genes based, which
can identify the important TFs according to the distribution of positive/negative genes. The second is absolute-based GSEA, which can
tackle the inter-gene-correlation within the target genes and reduce the false positives resulted from the gene permutation in GSEA. Finally,
algorithm outputs the TFs that both significant in two types of GSEA (\code{FDR <= cutoff}) and use the larger normalized enrichment score in two
types of GSEA as the \code{Reg_Score}.
}
\examples{
See RegSeq Tutorial for details.

}
\references{
Duanchen Sun and Zheng Xia (2018): RegSeq: A novel GSEA-based algorithm to perform the regulon analysis on RNA-Seq
data with small sample replicates.
}
