% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_Finding.R, R/Reg_Finding_original.R
\name{Twotailed}
\alias{Twotailed}
\alias{Twotailed}
\title{Twotailed: Reverse positive/negative target gene permuting GSEA.}
\usage{
Twotailed(regulon, tvalue, min, max, nPerm)

Twotailed(regulon, tvalue, min, max, nPerm)
}
\description{
Twotailed: Reverse positive/negative target gene permuting GSEA.

Twotailed: Reverse positive/negative target gene permuting GSEA.
}
